\name{RGG}
\alias{RGG}
\title{R GUI Generator}
\description{
  RGG is a GUI generator for R scripts. This package is meant only to be used with JGR.
  For more information please visit \url{http://rgg.r-forge.r-project.org/}
}
\usage{
  On loading it will automatically add a menu "RGG" in JGR main window. Use the submenu "load" to load an R-GUI.
}